feat: Introduce an RFD process.
This introduces a new Request for Discussion (RFD) process for
considering changes to the CVE Record Format. The details of
the proposal are in the commit.

This commit includes both the first RFD, proposing the
introduction an RFD process and modeling what the process
looks like, as well as a template for future RFDs to reuse.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>