feat: Base the Software ID RFD on the affected array
This rewrites the core content of the RFD to base the
proposed new fields on the `affected` array instead of basing
them on the `cpeApplicability` object as the prior version of
the RFD did. The motivation and outcomes are generally unchanged,
but the specifics of the proposed edits are now different.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>