feat: Rewrite 'Compatibility and Migration' in RFD template
Based on conversations with Matt Power, I've amended the text
of the "Compatibility and Migration" section of the RFD template
in a couple of ways:

1. Removed explicit reference to SchemaVer as the versioning scheme
   of choice for the CVE Record Format.
2. Introduced a set of questions which must specifically be
   answered in any submitted RFD.

The key goal here is to balance the need for a high level of rigor
around compatibility and migration in CVE because it is a
systemically-important and large multi-stakeholder system with the
need to not be overly prescriptive in the RFD template in ways
which may in fact reduce rigor by over-specializing RFD requirements
on today's considerations for what the important questions are.

The questions included are specific, and intended to capture key
concerns about forward compatibility and the impact of changes
on CVE consumers particularly.

Separately, references to SchemaVer were removed because, while
it is my understanding that SchemaVer is the version scheme of
choice for the CVE Record Format, that is not currently codified
and itself likely ought to go through an RFD process to firmly
resolve. In fact, Matt Power has raised concerns that SchemaVer
may be insufficiently expressive for the versioning constraints
under which CVE operates, and I'd rather not attempt to resolve
that question in the context of this process-focused RFD.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>