chore: Update back-compact / forward-compat defs
Amend the definitions of "backward compatibility" and
"forward compatibility" in the "Compatibility and Migration"
section of the RFD template to more clearly explain
requirements for RFD writers.

In the future we'd like this section to be more rigorous,
and ideally to provide a detailed breakdown of what kinds
of changes in the schema are considered "breaking," but
the versioning rules for the CVE Record Format aren't
sufficiently defined yet.

I recommend developing an RFD to solidify the versioning
rules, and then amending the template to require greater
rigor around compatibility in this section.

Signed-off-by: Andrew Lilley Brinker <abrinker@mitre.org>