feat: Make RFD *only* about Package URLs
After feedback from the CVE Board, this RFD is being split into two,
with this part covering *only* the addition of support for Package
URLs, and the other part (to be opened as a separate RFD) covering
support for OmniBOR Artifact IDs.

This commit does the split, and makes no edits besides removing the
OmniBOR pieces and smoothing out the remaining language where
necessary.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>

feat: ensure new IDs can't fulfill "identifier-like" requirement

Per discussion in the QWG, this amends the RFD to clarify that the new
identifier fields being proposed are not able to fulfill the "identifier-like"
requirement in the `product` object inside the `affected` array. While this
may be changed in the future, for today it is the easiest path forward for
CVE data consumers, who could adopt the new fields if _desirable_ but would
not be obligated to do so.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>