feat: ensure new IDs can't fulfill "identifier-like" requirement
Per discussion in the QWG, this amends the RFD to clarify that the new
identifier fields being proposed are not able to fulfill the "identifier-like"
requirement in the `product` object inside the `affected` array. While this
may be changed in the future, for today it is the easiest path forward for
CVE data consumers, who could adopt the new fields if _desirable_ but would
not be obligated to do so.

Signed-off-by: Andrew Lilley Brinker <alilleybrinker@gmail.com>